Consolidate date constants to core/platform
Created centralized DateConstants.kt in core/platform/date with:
- DAYS_IN_WEEK
- MONTHS_IN_QUARTER
- MONTHS_IN_YEAR
- QUARTERS_IN_YEAR
- FIRST_DAY_OF_MONTH

This eliminates duplicate private constants that were scattered across
habits/domain, habits/presentation, and main modules.
