+++ core/platform/src/commonMain/kotlin/space/be1ski/vibits/core/platform/date/DateConstants.kt
@@ -1,0 +1,7 @@
+package space.be1ski.vibits.core.platform.date
+
+const val DAYS_IN_WEEK = 7
+const val MONTHS_IN_QUARTER = 3
+const val MONTHS_IN_YEAR = 12
+const val QUARTERS_IN_YEAR = 4
+const val FIRST_DAY_OF_MONTH = 1
