+++ feature/habits/domain/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/domain/usecase/DateUtils.kt
@@ -5,8 +5,8 @@ import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
 import kotlinx.datetime.minus
+import space.be1ski.vibits.core.platform.date.MONTHS_IN_QUARTER
 
-private const val MONTHS_IN_QUARTER = 3
 private const val FIRST_QUARTER_INDEX = 1
 
 /**
