--- feature/main/src/commonMain/kotlin/space/be1ski/vibits/feature/main/domain/model/TimeConstants.kt
@@ -1,6 +1,0 @@
-package space.be1ski.vibits.feature.main.domain.model
-
-internal const val FIRST_DAY_OF_MONTH = 1
-internal const val DAYS_IN_WEEK = 7
-internal const val MONTHS_PER_QUARTER = 3
-internal const val MONTHS_IN_YEAR = 12
