+++ feature/main/src/commonMain/kotlin/space/be1ski/vibits/feature/main/domain/usecase/GetActivityRangeStartDateUseCase.kt
@@ -2,9 +2,9 @@ package space.be1ski.vibits.feature.main.domain.usecase
 
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
+import space.be1ski.vibits.core.platform.date.FIRST_DAY_OF_MONTH
+import space.be1ski.vibits.core.platform.date.MONTHS_IN_QUARTER
 import space.be1ski.vibits.feature.habits.domain.model.ActivityRange
-import space.be1ski.vibits.feature.main.domain.model.FIRST_DAY_OF_MONTH
-import space.be1ski.vibits.feature.main.domain.model.MONTHS_PER_QUARTER
 
 /**
  * Returns the first day of the given activity range.
@@ -15,7 +15,7 @@ object GetActivityRangeStartDateUseCase {
       is ActivityRange.Week -> range.startDate
       is ActivityRange.Month -> LocalDate(range.year, range.month, FIRST_DAY_OF_MONTH)
       is ActivityRange.Quarter -> {
-        val month = Month((range.index - 1) * MONTHS_PER_QUARTER + 1)
+        val month = Month((range.index - 1) * MONTHS_IN_QUARTER + 1)
         LocalDate(range.year, month, FIRST_DAY_OF_MONTH)
       }
       is ActivityRange.Year -> LocalDate(range.year, Month.JANUARY, FIRST_DAY_OF_MONTH)
